Change Element1D class nodal behaviour
The Element1D class works a bit differently from the rest of the code right now, where it uses a table of nodes instead of a dictionnary of nodes currently. This is a first update to the Element1D class in order to slowly move it from a table driven class to a dictionary driven class.