Reset Element2D Class as a Element1D clone
The old (current) implementation of Element2D was originally developped on the ideas I first had about approaching 2D elements. Now that I have completed 1D elements in a more structured way, I will be merging it's structure and functionalities over to the 2D space.

Some corrections to the code was also made to remove a debugging print and fix some typos in the introduction note.