Refine Triangular(Element2D) Class and Children Classes
I tried to refine the triangular classes to provide a better interaction between parent and child classes. Much of the development focus, from here on, will be to try to minimize the number of repeated methods and include them as high in the Classes tree as possible, allowing for reusability down-to child of child classes.