Add Element2D Class methods for solving B_e, update to V0.1.1
This closes #3

The version of the software has been updated to v0.1.1 with the addition of some methods to the Element2D Class.

These Element2D methods ultimately allow the calculation of  shape functions for any quad or triangular element. This includes B_e matrices in both the reference element and the actual element.