Convert all Numpy array calculation to Sympy Matrices
Since a lot of the calculation of inverses and transposes of matrices is done with expression within them (eta*xi), numpy has a lot of trouble dealing with them correctly as it is trying to convert them to Float. Sympy on the other hand transforms the matrices as expected without trouble.