Replace Matrix.dot() with a simple Matrix multiplication (M*M)
The dot method for Sympy Matrices has a strange method of calculating the dot product of two matrices.  Especially when column or line matrices are used. A simple fix is to use the intended multiplication in Sympy. 

Fix provide_de() which assumed a column matrix was going to be input. I changed this because I assume people using the method will understand that a line Matrix needs to be given.

I will change this in the future to provide a Transpose of the Matrix if needed.