Add index argument to pre-defined elements
Pre-defined elements now have an index number associated with them. 

This index is the element number and will be used instead of the element number that has been used in the past. This is due mostly to the possibility of multiple different elements being created in a same mesh and I am scared the implementation will hold in its current form.