Add Element2D __str__ definition to allow printing
All pre-defined element classes now have a __name__ that is simple to understand to the outside user.

In order to follow the change introduced a couple of commits ago on the indexing of elements, all instances of Element2D class and child classes will require an index to be provided for easier solving later on.