Change integrate_domain() method behaviour
Integrate domain was done at the element_type level. This has been changed so that it is done at the Element2D level with an element_type parser. It saved a few lines of code and now automatically assigns ref_element limits based on the element_type (e_type).

It behaviour also changed from integrating different type of functions (Ne_ref or Ne_ref * Ne_ref). For clarity in the formulation, this multiplication is done at the method call level (integrate_domaine(Ne_ref * Ne_ref)) which will improve readability and decrease the scope of this method.